Firm Integrations review * User should implicitly authorize application * Installation from list of available integrations * Fix gate check to gate auth * API methods for firm integrations managements was deleted

+++ app/Http/Controllers/Web/FirmIntegrationsController.php
@@ -7,16 +7,14 @@ namespace App\Http\Controllers\Web;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreFirmIntegrationRequest;
 use App\Http\Requests\UpdateFirmIntegrationRequest;
-use App\Http\Resources\FirmIntegrationResource;
-use App\Http\Resources\FirmResource;
-use App\Http\Resources\IntegrationResource;
 use App\Managers\FirmIntegrationsManager;
 use App\Models\Firm;
 use App\Models\FirmIntegration;
-use App\Models\User;
+use App\Models\Integration;
+use App\Models\Token;
 use App\Repositories\IntegrationsRepository;
 use Illuminate\Http\RedirectResponse;
-use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Redirect;
 use Inertia\Response;
@@ -24,6 +22,8 @@ use function abort_if;
 use function compact;
 use function data_get;
 use function inertia;
+use function json_encode;
+use const JSON_THROW_ON_ERROR;
 
 class FirmIntegrationsController extends Controller
 {
@@ -35,69 +35,31 @@ class FirmIntegrationsController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return Response
-     */
-    public function index(Firm $firm, Request $request, IntegrationsRepository $repository): Response
-    {
-        FirmResource::withoutWrapping();
-
-        /** @var User $user */
-        $user = $request->user();
-
-        return inertia('FirmIntegrations', [
-            'firm' => FirmResource::make($firm),
-            'firm_integrations' => FirmIntegrationResource::collection(
-                $firm
-                    ->integrationsInstalls()
-                    ->with(
-                        'integration',
-                        // fixme: Why?
-                        'firm',
-                    )
-                    ->paginate()
-            ),
-            'integrations' => IntegrationResource::collection(
-                $repository->getAvailableFor($user->id)->paginate()
-            ),
-        ]);
-    }
-
-    /**
-     * Show the form for creating a new resource.
+     * @param Firm $firm
      *
      * @return Response
      */
-    public function create(Request $request, Firm $firm, IntegrationsRepository $repository): Response
+    public function index(Firm $firm, IntegrationsRepository $repository): Response
     {
-        Gate::check('manage-integrations', $firm);
-
-        FirmResource::withoutWrapping();
-        FirmIntegrationResource::withoutWrapping();
-
-        /** @var User $user */
-        $user = $request->user();
-
-        return inertia('FirmIntegrationEdit', [
-            'firm' => FirmResource::make($firm),
-            'firm_integration' => FirmIntegrationResource::make(FirmIntegration::make([
-                // todo: place default values here
-            ])),
-            'integrations' => IntegrationResource::collection(
-                $repository->getAvailableFor($user->id)->paginate()
-            ),
+        return inertia('Firms/FirmIntegrations', [
+            'firm' => $firm,
+            'integrations_installs' => $firm->integrationsInstalls->loadMissing('integration'),
+            'integrations' => $repository->getAvailableFor(Auth::id())->get(),
         ]);
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  StoreFirmIntegrationRequest $request
+     * @param Firm $firm
+     * @param StoreFirmIntegrationRequest $request
+     * @param FirmIntegrationsManager $manager
      *
      * @return RedirectResponse
      */
     public function store(Firm $firm, StoreFirmIntegrationRequest $request, FirmIntegrationsManager $manager): RedirectResponse
     {
-        Gate::check('manage-integrations', $firm);
+        Gate::authorize('manage-integrations', $firm);
 
         $firmIntegration = $manager->install($firm->id, data_get($request->validated(), 'integration_id'));
 
@@ -117,44 +79,45 @@ class FirmIntegrationsController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  FirmIntegration  $firmIntegration
+     * @param Firm $firm
+     * @param FirmIntegration $firmIntegration
      *
      * @return Response
      */
-    public function edit(Request $request, Firm $firm, FirmIntegration $firmIntegration, IntegrationsRepository $repository): Response
+    public function edit(Firm $firm, FirmIntegration $firmIntegration): Response
     {
-        Gate::check('manage-integrations', $firm);
+        Gate::authorize('manage-integrations', $firm);
 
-        FirmResource::withoutWrapping();
-        FirmIntegrationResource::withoutWrapping();
+        $integration = $firmIntegration->integration;
 
-        /** @var User $user */
-        $user = $request->user();
-
-        return inertia('FirmIntegrationEdit', [
+        return inertia('Firms/FirmIntegrationEdit', [
             'firm' => $firm,
-            'firm_integration' => FirmIntegrationResource::make($firmIntegration),
-            'integrations' => IntegrationResource::collection(
-                $repository
-                    ->getAvailableFor($user->id)
-                    ->whereId($firmIntegration->integration_id)
-                    ->with('client')
-                    ->paginate()
-            ),
+            'install' => $firmIntegration,
+            'integration' => $integration,
+            'tokens' => $integration->o_auth2_client_id
+                ? Token::whereUserId(Auth::id())
+                ->where('revoked', false)
+                ->where('client_id', $integration->o_auth2_client_id)
+                ->get()
+                ->whereFirm($firm->id)
+                ->loadMissing('client')
+                ->values()
+                : [],
         ]);
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  UpdateFirmIntegrationRequest $request
-     * @param  FirmIntegration  $firmIntegration
+     * @param Firm $firm
+     * @param UpdateFirmIntegrationRequest $request
+     * @param FirmIntegration $firmIntegration
      *
      * @return RedirectResponse
      */
     public function update(Firm $firm, UpdateFirmIntegrationRequest $request, FirmIntegration $firmIntegration): RedirectResponse
     {
-        Gate::check('manage-integrations', $firm);
+        Gate::authorize('manage-integrations', $firm);
 
         $firmIntegration->update($request->validated());
 
@@ -168,16 +131,44 @@ class FirmIntegrationsController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  FirmIntegration  $firmIntegration
+     * @param Firm $firm
+     * @param FirmIntegration $firmIntegration
      *
      * @return RedirectResponse
      */
     public function destroy(Firm $firm, FirmIntegration $firmIntegration): RedirectResponse
     {
-        Gate::check('manage-integrations', $firm);
+        Gate::authorize('manage-integrations', $firm);
 
         $firmIntegration->delete();
 
         return Redirect::route('firms.firm_integrations.index', compact('firm'), 303);
+    }
+
+    public function auth(Firm $firm, FirmIntegration $firmIntegration): RedirectResponse
+    {
+        $integration = $firmIntegration->integration;
+        $client = $integration->client;
+
+        //if ($integration->auth === Integration::AUTH_OAUTH2) {
+        // @TODO: Another auth methods
+        //}
+
+        return Redirect::route(
+            'passport.authorizations.authorize',
+            [
+                'client_id' => $client->id,
+                'redirect_uri' => $client->redirect,
+                'response_type' => 'code',
+                'scope' => $integration->oauth2ScopesFor($firm->id),
+                'state' => json_encode([
+                    'client_id' => $client->id,
+                    'interface' => null,
+                    'user_id' => Auth::id(),
+                    'firm_id' => $firm->id,
+                ], JSON_THROW_ON_ERROR),
+            ],
+            303
+        );
     }
 }
